+++ Projects/Modules/Documentarian/Source/Public/Functions/General/Remove-Metadata.ps1
@@ -15,18 +15,24 @@ function Remove-Metadata {
     )
 
     foreach ($file in (Get-ChildItem -Path $Path -Recurse:$Recurse)) {
-        $file.name
-        $metadata = Get-Metadata -Path $file
-        $mdtext = Get-ContentWithoutHeader -Path $file
+
+        $doc = Get-Document -Path $file
 
         foreach ($key in $KeyName) {
-            if ($metadata.ContainsKey($key)) {
+            if ($doc.FrontMatter.ContainsKey($key)) {
                 $metadata.Remove($key)
             }
         }
-
-        Set-Content -Value (hash2yaml $metadata) -Path $file -Force -Encoding utf8
-        Add-Content -Value $mdtext -Path $file -Encoding utf8
+        $header = @(
+            '---'
+            [environment]::NewLine
+            $doc.FrontMatter | ConvertTo-Yaml
+            [environment]::NewLine
+            '---'
+        ) -join ''
+        Set-Content -Value $header  -Path $file -Force -Encoding utf8
+        Add-Content -Value $doc.Body -Path $file -Encoding utf8
+        $file
     }
 
 }